+++ lib/app/screens/order_screen/order_screen.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 
 import '../../utilities/colors.dart';
 import '../../utilities/text_styles.dart';
